PYTHON-6022 Consolidate all non-IO code in AsyncCursor/Cursor into _AgnosticCursor
